Extract WorkoutExecutionState enum to separate file
Move WorkoutExecutionState enum from WorkoutExecutionView.swift to its own file
in Models directory for better code organization and reusability.
