+++ WorkoutTimer/Models/WorkoutExecutionState.swift
@@ -1,0 +1,17 @@
+//
+//  WorkoutExecutionState.swift
+//  WorkoutTimer
+//
+//  Enum representing the state of workout execution.
+//
+
+import Foundation
+
+enum WorkoutExecutionState: Equatable {
+    case ready
+    case countdown
+    case running
+    case resting
+    case paused
+    case completed
+}
